2017-05-12 Erstellen der Modus und Grenzen klasse

+++ Code/src/rechnungen/Grenzen.java
@@ -1,10 +1,10 @@
 package rechnungen;
 
 public class Grenzen{
-    private int minZahl1;
-    private int minZahl2;
-    private int maxZahl1;
-    private int maxZahl2;
+    private double minZahl1;
+    private double minZahl2;
+    private double maxZahl1;
+    private double maxZahl2;
     private int kommaZahl1;
     private int kommaZahl2;
 }
